@@ -6,13 +6,12 @@ maintainers = [
   { name = "ViktorSky", email = "viktorbotsprojects@gmail.com" },
 ]
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = "MIT"
 license-files = ["LICENSE"]
 classifiers = [
   "Operating System :: OS Independent",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
@@ -50,7 +49,7 @@ optional-dependencies.dev = { file = "requirements-dev.txt" }
 "*" = ["py.typed"]
 
 [tool.pyright]
-pythonVersion = "3.8"
+pythonVersion = "3.9"
 typeCheckingMode = "strict"
 strictListInference = true
 strictDictionaryInference = true
