@@ -3,7 +3,7 @@ import os
 import threading
 import time
 from collections.abc import Iterator
-from typing import Any, Dict, Optional
+from typing import Any, Optional
 
 import diskcache
 
@@ -345,7 +345,7 @@ class Bot(socket.WSClient, global_client.Global):
         """
         return f"Bot: Prefix='{self.command_prefix}', Community ID={self.community_id}, Device ID='{self.device_id}'"
 
-    def __iter__(self) -> "Iterator[str]":
+    def __iter__(self) -> Iterator[str]:
         """
         Allows iteration over the Bot object.
 
@@ -665,7 +665,7 @@ class Bot(socket.WSClient, global_client.Global):
         password: Optional[str] = None,
         secret: Optional[str] = None,
         device_id: Optional[str] = None,
-    ) -> Dict[str, Any]:
+    ) -> dict[str, Any]:
         """
         Authenticates the bot with the provided email and password.
 
@@ -710,7 +710,7 @@ class Bot(socket.WSClient, global_client.Global):
         secret: Optional[str] = None,
         device_id: Optional[str] = None,
         use_cache: bool = True,
-    ) -> Dict[str, Any]:
+    ) -> dict[str, Any]:
         """
         Authenticates the user with the provided email and password.
 
@@ -791,7 +791,7 @@ class Bot(socket.WSClient, global_client.Global):
         sid: Optional[str] = None,
         device_id: Optional[str] = None,
         use_cache: bool = True,
-    ) -> Dict[str, Any]:
+    ) -> dict[str, Any]:
         """
         Logs in to the client using the provided credentials.
 
@@ -837,7 +837,7 @@ class Bot(socket.WSClient, global_client.Global):
             raise entities.LoginFailed
         return self._run(response)
 
-    def _run(self, response: Dict[str, Any]) -> Dict[str, Any]:
+    def _run(self, response: dict[str, Any]) -> dict[str, Any]:
         """
         Processes the response from a successful login attempt and sets up the authenticated client.
 
@@ -930,7 +930,7 @@ class Bot(socket.WSClient, global_client.Global):
             self._debug = False
             console.Console(self).fetch_menu()
 
-    def fetch_account(self) -> Dict[str, Any]:
+    def fetch_account(self) -> dict[str, Any]:
         """
         Fetches the account information for the authenticated user.
 
