@@ -1,5 +1,5 @@
 from collections.abc import Iterator
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Optional, Union
 
 from pymino.ext.entities import userprofile
 
@@ -13,8 +13,8 @@ __all__ = (
 
 
 class CommentStickerCollectionSummary:
-    def __init__(self, data: Dict[str, Any]):
-        self.data: Dict[str, Any] = data.get("stickerCollectionSummary", data) or {}
+    def __init__(self, data: dict[str, Any]):
+        self.data: dict[str, Any] = data.get("stickerCollectionSummary", data) or {}
 
     @property
     def status(self) -> int:
@@ -78,8 +78,8 @@ class CommentStickerCollectionSummary:
 
 
 class CommentSticker:
-    def __init__(self, data: Dict[str, Any]):
-        self.data: Dict[str, Any] = data.get("sticker", data) or {}
+    def __init__(self, data: dict[str, Any]):
+        self.data: dict[str, Any] = data.get("sticker", data) or {}
 
     @property
     def status(self) -> int:
@@ -133,8 +133,8 @@ class CommentSticker:
 
 
 class CommentExtensions:
-    def __init__(self, data: Dict[str, Any]):
-        self.data: Dict[str, Any] = data.get("extensions", data) or {}
+    def __init__(self, data: dict[str, Any]):
+        self.data: dict[str, Any] = data.get("extensions", data) or {}
 
     @property
     def sticker(self) -> CommentSticker:
@@ -155,11 +155,11 @@ class Comment:
     :type data: Union[dict, str]
     """
 
-    def __init__(self, data: Dict[str, Any]) -> None:
+    def __init__(self, data: dict[str, Any]) -> None:
         self.data = data
 
     @property
-    def author(self) -> Dict[str, Any]:
+    def author(self) -> dict[str, Any]:
         """Returns the author of the comment."""
         return self.data.get("author") or {}
 
@@ -184,9 +184,9 @@ class Comment:
         return CommentExtensions(self.data.get("extensions") or {})
 
     @property
-    def mediaList(self) -> List[Tuple[int, str]]:
+    def mediaList(self) -> list[tuple[int, str]]:
         """Returns the media list of the comment."""
-        mediaList: List[List[Any]] = self.data.get("mediaList") or []
+        mediaList: list[list[Any]] = self.data.get("mediaList") or []
         return [(m[0], m[1]) for m in mediaList]
 
     @property
@@ -249,7 +249,7 @@ class Comment:
         """Returns the sum of votes for the comment."""
         return self.data.get("votesSum", 0)
 
-    def json(self) -> Dict[str, Any]:
+    def json(self) -> dict[str, Any]:
         """Returns the raw data of the comment."""
         return self.data
 
@@ -262,7 +262,7 @@ class CommentList:
     :type data: dict
     """
 
-    def __init__(self, data: Union[List[Dict[str, Any]], Dict[str, Any]]):
+    def __init__(self, data: Union[list[dict[str, Any]], dict[str, Any]]):
         if not isinstance(data, list):
             data = list(data.get("commentList", data) or [])
         self.data = data
@@ -277,85 +277,85 @@ class CommentList:
         return userprofile.UserProfileList([comment.author for comment in self])
 
     @property
-    def commentId(self) -> List[str]:
+    def commentId(self) -> list[str]:
         """Returns a list of comment IDs in the comment list."""
         return [comment.commentId for comment in self]
 
     @property
-    def content(self) -> List[Optional[str]]:
+    def content(self) -> list[Optional[str]]:
         """Returns a list of comment IDs in the comment list."""
         return [comment.content for comment in self]
 
     @property
-    def createdTime(self) -> List[str]:
+    def createdTime(self) -> list[str]:
         """Returns a list of creation times of the comments in the comment list."""
         return [comment.createdTime for comment in self]
 
     @property
-    def extensions(self) -> List[CommentExtensions]:
+    def extensions(self) -> list[CommentExtensions]:
         """Returns a list of extensions of the comments in the comment list."""
         return [comment.extensions for comment in self]
 
     @property
-    def mediaList(self) -> List[List[Tuple[int, str]]]:
+    def mediaList(self) -> list[list[tuple[int, str]]]:
         """Returns a list of media lists of the comments in the comment list."""
         return [comment.mediaList for comment in self]
 
     @property
-    def modifiedTime(self) -> List[Optional[str]]:
+    def modifiedTime(self) -> list[Optional[str]]:
         """Returns a list of modification times of the comments in the comment list."""
         return [comment.modifiedTime for comment in self]
 
     @property
-    def ndcId(self) -> List[int]:
+    def ndcId(self) -> list[int]:
         """Returns a list of NDC IDs of the comments in the comment list."""
         return [comment.comId for comment in self]
 
     @property
-    def comId(self) -> List[int]:
+    def comId(self) -> list[int]:
         """Returns a list of NDC IDs of the comments in the comment list."""
         return [comment.comId for comment in self]
 
     @property
-    def parentId(self) -> List[str]:
+    def parentId(self) -> list[str]:
         """Returns a list of parent IDs of the comments in the comment list."""
         return [comment.parentId for comment in self]
 
     @property
-    def parentNdcId(self) -> List[int]:
+    def parentNdcId(self) -> list[int]:
         """Returns a list of parent NDC IDs of the comments in the comment list."""
         return [comment.parentNdcId for comment in self]
 
     @property
-    def parentComId(self) -> List[int]:
+    def parentComId(self) -> list[int]:
         """Returns a list of parent community IDs of the comments in the comment list."""
         return [comment.parentComId for comment in self]
 
     @property
-    def parentType(self) -> List[int]:
+    def parentType(self) -> list[int]:
         """Returns a list of parent types of the comments in the comment list."""
         return [comment.parentType for comment in self]
 
     @property
-    def subcommentsCount(self) -> List[int]:
+    def subcommentsCount(self) -> list[int]:
         """Returns a list of subcomment counts for the comments in the comment list."""
         return [comment.subcommentsCount for comment in self]
 
     @property
-    def type(self) -> List[int]:
+    def type(self) -> list[int]:
         """Returns a list of types of the comments in the comment list."""
         return [comment.type for comment in self]
 
     @property
-    def votedValue(self) -> List[int]:
+    def votedValue(self) -> list[int]:
         """Returns a list of voted values of the comments in the comment list."""
         return [comment.votedValue for comment in self]
 
     @property
-    def votesSum(self) -> List[int]:
+    def votesSum(self) -> list[int]:
         """Returns a list of vote sums for the comments in the comment list."""
         return [comment.votesSum for comment in self]
 
-    def json(self) -> List[Dict[str, Any]]:
+    def json(self) -> list[dict[str, Any]]:
         """Returns the raw data of the comment list."""
         return self.data
